chore(llc & llm): types for live-common/ble
