+++ libs/ledger-live-common/src/ble/hooks/useBleDevicesScanning.ts
@@ -2,58 +2,21 @@ import { useState, useEffect, useRef } from "react";
 import { Observable, of } from "rxjs";
 import { concatAll } from "rxjs/operators";
 import { getInfosForServiceUuid } from "@ledgerhq/devices";
-import type { DeviceModel, DeviceModelId } from "@ledgerhq/devices";
+import type { DeviceModelId } from "@ledgerhq/devices";
 import type {
   Observer as TransportObserver,
   Subscription as TransportSubscription,
   DescriptorEvent,
   DescriptorEventType,
 } from "@ledgerhq/hw-transport";
+import { TransportBleDevice, ScannedDevice, BleError } from "../types";
 
-// Inspired by Device class from react-native-ble-plx
-// Should be exported from somewhere else
-export type TransportBleDevice = {
-  // Device identifier: MAC address on Android and UUID on iOS.
-  id: string;
-  // Device name
-  name: string | null;
-  // User friendly name of device.
-  localName: string | null;
-  // Current Received Signal Strength Indication of device
-  rssi: number | null;
-  // Current Maximum Transmission Unit for this device.
-  // When device is not connected default value of 23 is used.
-  mtu: number;
-  // List of available services visible during scanning.
-  serviceUUIDs: string[] | null;
-  // Allows any other properties
-  [otherOptions: string]: unknown;
-};
-
-// Inspired by BleError class from react-native-ble-plx
-// BleError should be an error class which is guaranteed to be thrown by all functions
-// by our different implementations of Transport or at least of ble implementation of Transport
-// TODO: BleError as a class
-export type BleError = Error & {
-  // Platform independent error code.
-  // It is defined as an enum named BleErrorCode in react-native-ble-plx
-  // We should have our own mapping
-  errorCode: number;
-  // Allows any other properties
-  [otherOptions: string]: unknown;
-};
-
-export type ScannedDevice = {
-  deviceId: string;
-  deviceName: string;
-  bleRssi: number | null;
-  deviceModel: DeviceModel;
-};
+export type ScanningBleError = BleError | null;
 
 export type UseBleDevicesScanningResult = {
   scannedDevices: ScannedDevice[];
   scanningTimedOut: boolean;
-  scanningBleError: BleError | null;
+  scanningBleError: ScanningBleError;
 };
 
 export type UseBleDevicesScanningDependencies = {
@@ -87,9 +50,8 @@ export const useBleDevicesScanning = ({
 }: UseBleDevicesScanningDependencies &
   UseBleDevicesScanningOptions): UseBleDevicesScanningResult => {
   const [scanningTimedOut, setScanningTimedOut] = useState<boolean>(false);
-  const [scanningBleError, setScanningBleError] = useState<BleError | null>(
-    null
-  );
+  const [scanningBleError, setScanningBleError] =
+    useState<ScanningBleError>(null);
   const [scannedDevices, setScannedDevices] = useState<ScannedDevice[]>([]);
   // To check for duplicates. The ref will persist for the full lifetime of the component
   // in which the hook is called.
